requirements.txt
psycopg2-binary>=2.9.9
# Future additions (uncomment as needed)
# python-dotenv>=1.0.0